Bug 1480080 - Find in page return key not dismissing keyboard (#4125)
* Bug 1480080 - Find in page return key
Part 1 [code cleanup] move code verbatim out of BVC into BVC+FindInPage.swift

* Keyboard with a .search returnKeyType doesn't dismiss when return pressed. Handle this manually.
