+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -49,8 +49,8 @@ class BrowserViewController: UIViewController {
     fileprivate var screenshotHelper: ScreenshotHelper!
     fileprivate var homePanelIsInline = false
     fileprivate var searchLoader: SearchLoader?
-    fileprivate let alertStackView = UIStackView() // All content that appears above the footer should be added to this view. (Find In Page/SnackBars)
-    fileprivate var findInPageBar: FindInPageBar?
+    let alertStackView = UIStackView() // All content that appears above the footer should be added to this view. (Find In Page/SnackBars)
+    var findInPageBar: FindInPageBar?
 
     lazy var mailtoLinkHandler: MailtoLinkHandler = MailtoLinkHandler()
 
@@ -1063,41 +1063,6 @@ class BrowserViewController: UIViewController {
 
         present(controller, animated: true, completion: nil)
         LeanPlumClient.shared.track(event: .userSharedWebpage)
-    }
-
-    func updateFindInPageVisibility(visible: Bool, tab: Tab? = nil) {
-        if visible {
-            if findInPageBar == nil {
-                let findInPageBar = FindInPageBar()
-                self.findInPageBar = findInPageBar
-                findInPageBar.delegate = self
-                alertStackView.addArrangedSubview(findInPageBar)
-
-                findInPageBar.snp.makeConstraints { make in
-                    make.height.equalTo(UIConstants.ToolbarHeight)
-                    make.leading.trailing.equalTo(alertStackView)
-                }
-
-                updateViewConstraints()
-
-                // We make the find-in-page bar the first responder below, causing the keyboard delegates
-                // to fire. This, in turn, will animate the Find in Page container since we use the same
-                // delegate to slide the bar up and down with the keyboard. We don't want to animate the
-                // constraints added above, however, so force a layout now to prevent these constraints
-                // from being lumped in with the keyboard animation.
-                findInPageBar.layoutIfNeeded()
-            }
-
-            self.findInPageBar?.becomeFirstResponder()
-        } else if let findInPageBar = self.findInPageBar {
-            findInPageBar.endEditing(true)
-            let tab = tab ?? tabManager.selectedTab
-            guard let webView = tab?.webView else { return }
-            webView.evaluateJavaScript("__firefox__.findDone()", completionHandler: nil)
-            findInPageBar.removeFromSuperview()
-            self.findInPageBar = nil
-            updateViewConstraints()
-        }
     }
 
     @objc fileprivate func openSettings() {
@@ -2700,40 +2665,6 @@ extension BrowserViewController: Themeable {
         homePanelController?.applyTheme()
 
         (presentedViewController as? Themeable)?.applyTheme()
-    }
-}
-
-extension BrowserViewController: FindInPageBarDelegate, FindInPageHelperDelegate {
-    func findInPage(_ findInPage: FindInPageBar, didTextChange text: String) {
-        find(text, function: "find")
-    }
-
-    func findInPage(_ findInPage: FindInPageBar, didFindNextWithText text: String) {
-        findInPageBar?.endEditing(true)
-        find(text, function: "findNext")
-    }
-
-    func findInPage(_ findInPage: FindInPageBar, didFindPreviousWithText text: String) {
-        findInPageBar?.endEditing(true)
-        find(text, function: "findPrevious")
-    }
-
-    func findInPageDidPressClose(_ findInPage: FindInPageBar) {
-        updateFindInPageVisibility(visible: false)
-    }
-
-    fileprivate func find(_ text: String, function: String) {
-        guard let webView = tabManager.selectedTab?.webView else { return }
-        let escaped = text.replacingOccurrences(of: "\\", with: "\\\\").replacingOccurrences(of: "\"", with: "\\\"")
-        webView.evaluateJavaScript("__firefox__.\(function)(\"\(escaped)\")", completionHandler: nil)
-    }
-
-    func findInPageHelper(_ findInPageHelper: FindInPageHelper, didUpdateCurrentResult currentResult: Int) {
-        findInPageBar?.currentResult = currentResult
-    }
-
-    func findInPageHelper(_ findInPageHelper: FindInPageHelper, didUpdateTotalResults totalResults: Int) {
-        findInPageBar?.totalResults = totalResults
     }
 }
