+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+FindInPage.swift
@@ -1,0 +1,76 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import Shared
+
+extension BrowserViewController {
+    func updateFindInPageVisibility(visible: Bool, tab: Tab? = nil) {
+        if visible {
+            if findInPageBar == nil {
+                let findInPageBar = FindInPageBar()
+                self.findInPageBar = findInPageBar
+                findInPageBar.delegate = self
+                alertStackView.addArrangedSubview(findInPageBar)
+
+                findInPageBar.snp.makeConstraints { make in
+                    make.height.equalTo(UIConstants.ToolbarHeight)
+                    make.leading.trailing.equalTo(alertStackView)
+                }
+
+                updateViewConstraints()
+
+                // We make the find-in-page bar the first responder below, causing the keyboard delegates
+                // to fire. This, in turn, will animate the Find in Page container since we use the same
+                // delegate to slide the bar up and down with the keyboard. We don't want to animate the
+                // constraints added above, however, so force a layout now to prevent these constraints
+                // from being lumped in with the keyboard animation.
+                findInPageBar.layoutIfNeeded()
+            }
+
+            self.findInPageBar?.becomeFirstResponder()
+        } else if let findInPageBar = self.findInPageBar {
+            findInPageBar.endEditing(true)
+            let tab = tab ?? tabManager.selectedTab
+            guard let webView = tab?.webView else { return }
+            webView.evaluateJavaScript("__firefox__.findDone()", completionHandler: nil)
+            findInPageBar.removeFromSuperview()
+            self.findInPageBar = nil
+            updateViewConstraints()
+        }
+    }
+}
+
+extension BrowserViewController: FindInPageBarDelegate, FindInPageHelperDelegate {
+    func findInPage(_ findInPage: FindInPageBar, didTextChange text: String) {
+        find(text, function: "find")
+    }
+
+    func findInPage(_ findInPage: FindInPageBar, didFindNextWithText text: String) {
+        findInPageBar?.endEditing(true)
+        find(text, function: "findNext")
+    }
+
+    func findInPage(_ findInPage: FindInPageBar, didFindPreviousWithText text: String) {
+        findInPageBar?.endEditing(true)
+        find(text, function: "findPrevious")
+    }
+
+    func findInPageDidPressClose(_ findInPage: FindInPageBar) {
+        updateFindInPageVisibility(visible: false)
+    }
+
+    fileprivate func find(_ text: String, function: String) {
+        guard let webView = tabManager.selectedTab?.webView else { return }
+        let escaped = text.replacingOccurrences(of: "\\", with: "\\\\").replacingOccurrences(of: "\"", with: "\\\"")
+        webView.evaluateJavaScript("__firefox__.\(function)(\"\(escaped)\")", completionHandler: nil)
+    }
+
+    func findInPageHelper(_ findInPageHelper: FindInPageHelper, didUpdateCurrentResult currentResult: Int) {
+        findInPageBar?.currentResult = currentResult
+    }
+
+    func findInPageHelper(_ findInPageHelper: FindInPageHelper, didUpdateTotalResults totalResults: Int) {
+        findInPageBar?.totalResults = totalResults
+    }
+}
